refactor: extract getBaseUrl into shared module
Move the request-to-base-URL helper out of index.ts into src/url.ts so the
OAuth router can build the Google callback URL from it without creating a
circular import (index.ts <-> oauth.ts).
